requirements.txt
alembic==1.7.7
APScheduler==3.10.4
backports.zoneinfo==0.2.1
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
Flask==2.0.3
Flask-APScheduler==1.12.4
Flask-Migrate==4.0.1
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mysql-connector==2.2.9
protobuf==3.19.6
psutil==5.9.6
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.50
typing_extensions==4.1.1
tzdata==2023.3
tzlocal==4.2
urllib3==1.26.18
Werkzeug==2.0.3
zipp==3.6.0